Split raygui extras into separate files for easier handling. Made timeline more efficient.
