+++ src/CreateEAXReverbZoneDialog.cpp
@@ -2,7 +2,7 @@
 
 #include "raygui/raygui.h"
 
-#include "raygui/raygui_extras.h"
+#include "raygui/gui_valid_float.h"
 
 CreateEAXReverbZoneDialog::CreateEAXReverbZoneDialog() 
 {
