+++ src/CreateSoundProducerDialog.cpp
@@ -4,7 +4,7 @@
 
 #include "raygui/raygui.h"
 
-#include "raygui/raygui_extras.h"
+#include "raygui/gui_valid_float.h"
 
 #include <cstring>
