+++ src/CreateStandardReverbZoneDialog.cpp
@@ -4,7 +4,8 @@
 
 #include "raygui/raygui.h"
 
-#include "raygui/raygui_extras.h"
+#include "raygui/gui_valid_float.h"
+#include "raygui/gui_dropdown_listview.h"
 
 CreateStandardReverbZoneDialog::CreateStandardReverbZoneDialog() 
 {
