+++ src/EditMultipleEchoZonesDialog.cpp
@@ -4,7 +4,8 @@
 
 #define GUI_DROPDOWN_LISTVIEW
 
-#include "raygui/raygui_extras.h"
+#include "raygui/gui_valid_float.h"
+#include "raygui/gui_dropdown_listview.h"
 
 #include <string>
