+++ src/EditMultipleSoundProducersDialog.cpp
@@ -7,7 +7,8 @@
 #include <cstring>
 #include <string>
 
-#include "raygui/raygui_extras.h"
+#include "raygui/gui_valid_float.h"
+#include "raygui/gui_dropdown_listview.h"
 
 EditMultipleSoundProducersDialog::EditMultipleSoundProducersDialog(const std::string& title)
 {
