+++ src/raygui/gui_valid_float.h
@@ -195,250 +195,4 @@ bool GuiTextBox_ValidValueFloatSimple(Rectangle rect, int textSize, ValidFloatPa
 
 #endif
 
-//GuiListView with Dropdown box bound change
 
-#ifndef GUI_DROPDOWN_LISTVIEW_H
-#define GUI_DROPDOWN_LISTVIEW_H
-
-typedef struct 
-{
-	int itemsShowCount;
-	bool editMode;
-	bool valueChanged;
-	int scrollIndex;
-} DropDownListViewSettings;
-
-#ifdef __cplusplus
-extern "C" {            // Prevents name mangling of functions
-#endif
-
-DropDownListViewSettings InitDropDownListViewSettings(int itemsShowCount, bool editMode, bool valueChanged, int scrollIndex);
-
-//function to use guilistview when drop down box is selected.
-int Gui_Dropdown_ListView(Rectangle bounds, int itemsShowCount, int itemsCount, const char *text, int *scrollIndex, int active, bool* editMode, bool* valueChanged);
-
-//function to use guilistview when drop down box is selected.
-int Gui_Dropdown_ListView_Simple(DropDownListViewSettings* settings,Rectangle bounds, const char *text, int itemsCount, int active);
-
-#ifdef __cplusplus
-}
-#endif
-
-
-#endif
-
-#if defined(GUI_DROPDOWN_LISTVIEW)
-
-
-int Gui_Dropdown_ListView(Rectangle bounds, int itemsShowCount, int itemsCount, const char *text, int *scrollIndex, int active, bool* editMode, bool* valueChanged)
-{
-	
-	Rectangle boundsOpen = bounds;
-    boundsOpen.height = (itemsShowCount + 1)*(bounds.height + GuiGetStyle(DROPDOWNBOX, DROPDOWN_ITEMS_PADDING));
-		
-	Rectangle& current_bounds = bounds;
-		
-	// Update control
-    //--------------------------------------------------------------------
-    
-	Vector2 mousePoint = GetMousePosition();
-
-	if (*editMode)
-	{
-		// Check if mouse has been pressed or released outside limits
-		if (!CheckCollisionPointRec(mousePoint, boundsOpen))
-		{
-			if (IsMouseButtonPressed(MOUSE_LEFT_BUTTON) || IsMouseButtonReleased(MOUSE_LEFT_BUTTON))
-			{
-				*editMode = false;
-				*scrollIndex = active;
-			} 
-		}
-		
-	}
-	else
-	{
-		//check if mouse clicked on drop down box
-		if (CheckCollisionPointRec(mousePoint, bounds))
-		{
-			if (IsMouseButtonPressed(MOUSE_LEFT_BUTTON))
-			{
-				*editMode = true;
-			}
-		}
-	}
-
-	if (*editMode){current_bounds = boundsOpen;}
-	
-	int itemSelected = -1;
-	
-	
-	itemSelected = GuiListView( current_bounds, text, scrollIndex, active );
-		
-	
-	if(itemSelected != -1 && itemSelected < itemsCount && itemSelected != active)
-	{
-		*valueChanged = true;
-		*editMode = false;
-		*scrollIndex = itemSelected;
-	}
-	else
-	{
-		itemSelected = active;
-		*valueChanged = false;
-	}
-	
-	return itemSelected;
-	
-}
-
-DropDownListViewSettings InitDropDownListViewSettings(int itemsShowCount, bool editMode, bool valueChanged, int scrollIndex)
-{
-	DropDownListViewSettings settings = { 0 };
-	
-	settings.itemsShowCount = itemsShowCount;
-	settings.editMode = false;
-	settings.valueChanged = false;
-	settings.scrollIndex = 0;
-	
-	return settings;
-}
-
-int Gui_Dropdown_ListView_Simple(DropDownListViewSettings* settings,Rectangle bounds, const char *text, int itemsCount, int active)
-{
-	return Gui_Dropdown_ListView(bounds, settings->itemsShowCount, itemsCount, text, &settings->scrollIndex, active, &settings->editMode, &settings->valueChanged);
-}
-
-
-#endif
-
-
-//3d Object Timeline
-
-#ifndef GUI_3D_OBJECT_TIMELINE_H
-#define GUI_3D_OBJECT_TIMELINE_H
-
-
-typedef struct 
-{
-	bool editMode;
-	
-	
-	size_t max_num_frames;
-	
-	size_t* array_points_ptr; //array of points to render
-	
-	bool valueChanged;
-	bool addPoint;
-	bool showPropertiesBox;
-	
-	//frame selection variables, 
-	size_t current_timeline_frame;
-	bool frameSelected;
-	float frameDrawX;
-	
-} TimelineSettings;
-
-#ifdef __cplusplus
-extern "C" {            // Prevents name mangling of functions
-#endif
-
-
-TimelineSettings InitTimelineSettings(size_t max_num_frames, size_t* points_ptr);
-
-//Draws timeline 
-//returns if timeline was clicked on
-bool Gui_3dObject_Timeline(TimelineSettings* settings);
-
-#ifdef __cplusplus
-}
-#endif
-
-#endif
-
-#if defined(GUI_3D_OBJECT_TIMELINE)
-
-
-TimelineSettings InitTimelineSettings(size_t max_num_frames, size_t* points_ptr)
-{
-	TimelineSettings settings = { 0 };
-	
-	settings.array_points_ptr = points_ptr;
-	settings.max_num_frames = max_num_frames;
-	settings.frameSelected = false;
-}
-
-bool Gui_3dObject_Timeline(TimelineSettings* settings)
-{
-	
-//draw timeline
-//horizontal axis is the time
-//vertical axis is nothing
-
-//draw box with background color
-float screen_height = GetScreenHeight();
-Rectangle drawAreaRect = {0, 400 , GetScreenWidth(), screen_height};
-
-DrawRectangleRec(drawAreaRect, Fade(GRAY, 0.8f));
-
-float leftBound = 200;
-Rectangle mouseArea = {leftBound, 400 , GetScreenWidth(), screen_height};
-
-//draw cursor line where mouse pointer is
-
-Vector2 mousePoint = GetMousePosition();
-
-//select frame based on where mouse pointer is clicked on 
-if (CheckCollisionPointRec(mousePoint, mouseArea))
-{
-	float cursor_x = mousePoint.x;
-	if(cursor_x < leftBound){cursor_x = leftBound;}
-
-	Color cursor_color = BLACK;
-	if (IsMouseButtonPressed(MOUSE_LEFT_BUTTON))
-	{
-		settings->frameSelected = !settings->frameSelected;
-		
-		if(settings->frameSelected)
-		{
-			settings->frameDrawX = cursor_x;
-		}
-	}
-	
-	if(settings->frameSelected)
-	{
-		cursor_x = settings->frameDrawX;
-		cursor_color = YELLOW;
-	}
-	
-	DrawRectangle(cursor_x,400,2,screen_height,cursor_color);
-}
-else
-{
-	if(settings->frameSelected)
-	{
-		float cursor_x = settings->frameDrawX;
-		Color cursor_color = YELLOW;
-		DrawRectangle(cursor_x,400,2,screen_height,cursor_color);
-	}
-	
-	
-}
-
-
-
-
-//draw points
-
-//if mouse click on point in timeline and add point is false
-	//show properties box to true
-		
-//draw properties box
-	
-//if show properties box bool is true
-	
-	//show the contents of the point in timeline such as position and rotation.
-	
-}
-
-#endif
